feat(web): introduce AgendaSkeleton and restructure event components for better organization
