+++ packages/web/src/views/Day/components/Agenda/Agenda.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useRef } from "react";
 import { useDayEvents } from "@web/views/Day/data/day.data";
 import { useDateInView } from "@web/views/Day/hooks/navigation/useDateInView";
-import { AgendaAllDayEvent } from "./AgendaAllDayEvent/AgendaAllDayEvent";
-import { AgendaEvents } from "./AgendaEvents/AgendaEvents";
+import { AgendaEvents } from "./Events/AgendaEvent/AgendaEvents";
+import { AllDayAgendaEvents } from "./Events/AllDayAgendaEvent/AllDayAgendaEvents";
 import { TimeLabels } from "./TimeLabels/TimeLabels";
 
 export function Agenda() {
@@ -28,15 +28,7 @@ export function Agenda() {
       aria-label="calendar-agenda"
       className="bg-darkBlue-400 flex h-full min-w-xs flex-col"
     >
-      {/* All-day events section - above the scroll container */}
-      {allDayEvents.length > 0 && (
-        <div className="space-y-1 border-b border-gray-200 bg-white px-4 py-2">
-          {allDayEvents.map((event) => (
-            <AgendaAllDayEvent key={event._id} event={event} />
-          ))}
-        </div>
-      )}
-
+      <AllDayAgendaEvents allDayEvents={allDayEvents} />
       <div
         ref={scrollRef}
         className="relative flex flex-1 overflow-x-hidden overflow-y-auto"
